refactor: move conversations to separate component
